+++ CHANGES.rst
@@ -5,6 +5,9 @@ New Features
 
 - Added ``insert_frame`` method to modify the pipeline of a ``WCS`` object. [#299]
 
+- Added ``to_fits_tab`` method to generate FITS header and binary table
+  extension following FITS WCS ``-TAB`` convension. [#295]
+
 0.14.0 (2020-08-19)
 -------------------
 New Features